+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.3] - 2025-10-21
+
+- fix: Preserve participant text during raffle reset and improve text field initialization
+- feat: Implement reset and clear winners dialogs in raffle screen.
+
 ## [1.1.2] - 2025-09-30
 
 - Updated all platform versions to maintain consistency across stores